WIP: ui access progress
We should name our document an accessibility document to emphasize that
it's just another output of our UI, dedicated to accessibility.

Rather than opting for a semi automatic output from graphical centric
UI to the accessibility document, it should be a thoughtful effort to make
a document that's great to use for impaired users.