Attempt fixing the issue with UI stuck at end screen while timer running.
There are cases where the UI is showing the end screen but the noise sound
is still playing, and no celeberation (notification) has took place. This
indicates that the UI code believes the timer to be expired, but the timer
corountine hasn't actually been called.

I believe this is possible when the UI repainting is triggered just before
the timer expires, and the repainting happens when the timer has expired.

In that case the timer running the coroutine would be set to a huge
validity period, and basically never run, causing no celebration to take
place and the noise to keep running.

I believe this could be prevented if the notion of current time was owned
by the timer itself, so that the UI could never diverge.

This fix is to invalidate the painting immediately when the timer has
expired but is still considered active.